Address ChatMessage refactor follow-ups: dedupe type re-export, split styles, hoist constants
Five follow-ups for the ChatMessage decomposition (PR #1586) review:

- ChatMessage.tsx: collapse the duplicate `import type { TimelineEntry } from
  "./types"` + `export type { TimelineEntry } from "./types"` pair into a
  single import + local re-export. The previous pair compiled but read as an
  oversight: the import was redundant against the re-export-from form.

- ChatMessage.styles.ts: split the 1,070-line styles barrel into four
  co-located sibling modules:
    * ChatMessage.styles.ts (412 lines, container/avatar/bubble/indicators)
    * ChatMessageSourcePreview.styles.ts (203 lines, new)
    * ChatMessageTimeline.styles.ts (331 lines, includes the streaming
      ticker styles since they share the typeColor helper with the
      timeline row icons)
    * ChatMessageToolUsage.styles.ts (187 lines, new)
  Each consumer now imports its styles from its own sibling module rather
  than from a barrel. The unused `SourcesContainer` export (no in-tree
  consumers) was dropped.

- osLegalStyles.ts: lift the most-reused chat-bubble palette out of the
  styles file into the central design tokens. Adds chatSourceBlue/Hover,
  chatMessageTextAssistant/User, chatSourcePreviewText, chatCodeText,
  chatUsernameText, chatAvatarAssistantStart/End, plus a coolGray scale
  (400/500/600/700/800) for timeline & code-block text. Also adds two
  small alpha helpers — whiteAlpha() and blackAlpha() — to mirror the
  existing primaryBlueAlpha/chatSourceBlueAlpha pattern. Niche per-type
  timeline accents (#a855f7 thought-purple, #f97316 content-orange) stay
  inline in `typeColor` since they don't recur elsewhere.

- constants.ts: add SMALL_MOBILE_BREAKPOINT = 480 and use it together with
  the existing TABLET_BREAKPOINT = 768 in the chat styles, removing the
  hardcoded `@media (max-width: 480px)` and `@media (max-width: 768px)`
  numbers per the no-magic-numbers rule.

- types.ts: refresh the leading docstring so it reflects the new sibling
  layout (multiple co-located *.styles.ts modules) rather than the
  old single-barrel layout.

The .any-baseline.json `widgets_chat: 2 → key removed` change from PR #1586
was sanity-checked: `node frontend/scripts/check-any-baseline.js` confirms
zero `any`s under `components/widgets/chat/**` and the script omits any area
whose count is zero (`if (n === 0) continue`). The removal is correct.

Test plan:
- tsc --noEmit clean
- prettier --check clean across the chat dir + configuration files
- 12 unit tests in chatMessageUtils.test.ts pass
- check-any-baseline.js: current=451 baseline=451, no regression
