+++ frontend/src/assets/configurations/constants.ts
@@ -1,6 +1,10 @@
 import { OS_LEGAL_COLORS } from "./osLegalStyles";
 
 export const VERSION_TAG = "v3.0.0.rc1";
+// Small-mobile breakpoint - used by chat widget styles for the tightest
+// viewports (very small phones), where chat-bubble arrows are removed and
+// font sizes shrink. Distinct from MOBILE_VIEW_BREAKPOINT (600).
+export const SMALL_MOBILE_BREAKPOINT = 480;
 export const MOBILE_VIEW_BREAKPOINT = 600;
 // Tablet breakpoint - used for sidebar collapse behavior (larger than mobile)
 export const TABLET_BREAKPOINT = 768;
